+++ src/scripts/InstanceScripts/Instance_HallsOfLightning.cpp
@@ -378,7 +378,7 @@ class VolkhansAnvil : public CreatureAIScript
     VolkhansAnvil(Creature* pCreature) : CreatureAIScript(pCreature)
     {
         getCreature()->GetAIInterface()->SetAllowedToEnterCombat(false);
-        getCreature()->setUInt64Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
+        getCreature()->setUInt32Value(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE);
         setRooted(true);
     }
